+++ src/views/projectExplorer/ifsFile.ts
@@ -5,7 +5,7 @@
 import { ThemeIcon, TreeItemCollapsibleState, Uri, WorkspaceFolder } from "vscode";
 import * as path from "path";
 import { ProjectExplorerTreeItem } from "./projectTreeItem";
-import { ContextValue } from "../../typings";
+import { ContextValue } from "../../projectExplorerApi";
 
 export default class IFSFile extends ProjectExplorerTreeItem {
   static contextValue = ContextValue.ifsFile;
